portfolio_construction/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
nltk
matplotlib
seaborn
markdown2
yfinance
pandas
numpy
quantstats
pypfopt
requests
    